set images on certain containers only
