+++ pkg/model/deployment_test.go
@@ -60,11 +60,23 @@ func TestSpecs(t *testing.T) {
 
 }
 
+func TestSpecImagePullSecrets(t *testing.T) {
+	bs := *deploymentTestBackstage.DeepCopy()
+
+	bs.Spec.Application.ImagePullSecrets = nil //[]string{}
+
+	testObj := createBackstageTest(bs).withDefaultConfig(true).
+		addToDefaultConfig("deployment.yaml", "janus-deployment.yaml")
+
+	model, err := InitObjects(context.TODO(), bs, testObj.externalConfig, true, true, testObj.scheme)
+	assert.NoError(t, err)
+
+	assert.Equal(t, 0, len(model.backstageDeployment.deployment.Spec.Template.Spec.ImagePullSecrets))
+	//assert.Equal(t, "my-secret", model.backstageDeployment.deployment.Spec.Template.Spec.ImagePullSecrets[0].Name)
+
+}
+
 // It tests the overriding image feature
-// [GA] if we need this (and like this) feature
-// we need to think about simple template engine
-// for substitution env vars instead.
-// Janus image specific
 func TestOverrideBackstageImage(t *testing.T) {
 
 	bs := *deploymentTestBackstage.DeepCopy()
@@ -78,5 +90,6 @@ func TestOverrideBackstageImage(t *testing.T) {
 	assert.NoError(t, err)
 
 	assert.Equal(t, "dummy", model.backstageDeployment.container().Image)
+	assert.Equal(t, "dummy", model.backstageDeployment.podSpec().InitContainers[0].Image)
 
 }
